fastapi_project/app/requirements.txt
fastapi
uvicorn
databases
passlib[bcrypt]
python-dotenv
python-jose
fastapi_cache
PyJWT
jwt